Print basic info per generation

Print basic info per generation

callback_base_report (userdir, generation, pop_pool,
pareto, prob_matrix, current_time)

Arguments

userdir
the default directory used by `galgo()` to store files

generation
a number indicating the number of iterations of the galgo algorithm

pop_pool
a data.frame with the solution vectors, number of clusters and their ranking.

pareto
the solutions found by Galgo across all generations in the solution space

prob_matrix
a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns

current_time
an POSIXct object

Value

Nothing.

Examples

# load example dataset

library(breastCancerTRANSBIG)
data(transbig)
Train <- transbig
rm(transbig)

#> Warning: object 'transbig' not found

expression <- Biobase::exprs(Train)
clinical <- Biobase::pData(Train)
OS <- survival::Surv(time = clinical$t.rfs, event = clinical$e.rfs)
# We will use a reduced dataset for the example
expression <- expression[sample(1:nrow(expression), 100), ]

# Now we scale the expression matrix
expression <- t(scale(t(expression)))

# Run galgo with base_report_callback assigned to the report_callback
# hook-point
GSgalgoR::galgo(generations = 5,
population = 15,
prob_matrix = expression,
OS = OS,
report_callback = callback_base_report
)

#> Using CPU for computing pearson distance

#> TRUETRUETRUETRUETRUETRUETRUETRUEFALSETRUETRUETRUETRUETRUETRUE

#> 14100

#> Generation 1 Non-dominated solutions:

#>           k                      rnkIndex     CrowD
#> result.7  2 0.10888047  14.49866        1       Inf
#> result.9  5 0.05056947 185.49095        1 0.7214049
#> result.10 2 0.06498447 160.66696        1 1.3679277
#> result.14 7 0.02128337 224.89984        1       Inf

#> TRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUE

#> 15100

#> Generation 2 Non-dominated solutions:

#>            k                      rnkIndex     CrowD
#> result.7   2 0.10888047  14.49866        1       Inf
#> result.14  7 0.02128337 224.89984        1       Inf
#> result.10  2 0.09127924  97.48479        1 1.0479639
#> result.10  2 0.06498447 160.66696        1 0.7587684
#> result.9   5 0.05056947 185.49095        1 0.6560957
#> result.12 10 0.02257438 222.37007        1 0.4387219

#> TRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUE

#> 15100

#> Generation 3 Non-dominated solutions:

#>            k                      rnkIndex     CrowD
#> result.7   2 0.10888047  14.49866        1       Inf
#>           10 0.01637632 254.46413        1       Inf
#> result.10  2 0.09127924  97.48479        1 0.9732332
#> result.10  2 0.06498447 160.66696        1 0.7145353
#> result.9   5 0.05056947 185.49095        1 0.6258221
#> result.12 10 0.02257438 222.37007        1 0.4195039
#> result.14  7 0.02128337 224.89984        1 0.1827214

#> TRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUE

#> 15100

#> Generation 4 Non-dominated solutions:

#>            k                      rnkIndex     CrowD
#> result.7   2 0.10888047  14.49866        1       Inf
#>           10 0.01637632 254.46413        1       Inf
#> result.10  2 0.09127924  97.48479        1 0.9524847
#> result.10  2 0.06498447 160.66696        1 0.6964094
#> result.9   5 0.05056947 185.49095        1 0.6076431
#> result.12 10 0.02257438 222.37007        1 0.4070256
#> result.14  7 0.02128337 224.89984        1 0.1795228

#> TRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUETRUE

#> 15100

#> Generation 5 Non-dominated solutions:

#>            k                      rnkIndex     CrowD
#> result.7   2 0.10888047  14.49866        1       Inf
#>           10 0.01637632 254.46413        1       Inf
#> result.10  2 0.09127924  97.48479        1 0.9381609
#> result.10  2 0.06498447 160.66696        1 0.6797534
#>            7 0.04094213 227.00361        1 0.5511496
#> result.9   5 0.05056947 185.49095        1 0.4589268

#> An object of class "galgo.Obj"
#> Slot "Solutions":
#>           205766_at 38241_at 207078_at 210843_s_at 200957_s_at 219627_at
#> result.7          0        0         1           0           0         1
#>                   1        1         0           1           1         1
#> result.10         0        0         1           0           0         1
#> result.10         1        1         1           1           1         1
#>                   1        0         0           1           1         1
#> result.9          1        0         0           0           1         0
#>                   0        0         1           0           0         1
#>                   0        0         1           0           0         1
#> result.7          0        0         1           1           0         0
#>                   1        0         1           1           1         1
#>                   0        1         1           1           1         1
#> result.8          1        1         0           1           1         1
#>                   1        1         1           1           1         1
#> result.14         0        0         1           0           0         1
#> result.12         0        1         0           1           1         1
#>           215890_at 218495_at 212079_s_at 202717_s_at 221636_s_at 213443_at
#> result.7          0         1           0           0           0         0
#>                   1         1           1           1           1         1
#> result.10         0         1           0           0           0         0
#> result.10         1         1           1           1           1         1
#>                   1         0           0           0           0         1
#> result.9          0         0           0           0           0         0
#>                   0         1           0           0           0         0
#>                   0         1           0           0           0         0
#> result.7          0         0           0           0           1         0
#>                   1         0           0           0           1         0
#>                   0         1           1           1           0         1
#> result.8          1         1           1           1           1         1
#>                   1         1           1           1           1         1
#> result.14         0         0           1           1           0         0
#> result.12         1         1           1           1           1         1
#>           204646_at 219635_at 212306_at 216050_at 213462_at 216458_at 203845_at
#> result.7          0         1         0         0         0         1         1
#>                   1         1         1         1         1         1         1
#> result.10         0         1         0         0         0         1         1
#> result.10         1         1         1         1         1         1         1
#>                   0         1         1         0         1         1         0
#> result.9          0         1         1         0         1         1         0
#>                   0         1         0         0         0         1         1
#>                   0         1         0         0         0         1         1
#> result.7          0         1         0         1         0         1         1
#>                   0         1         0         0         0         1         1
#>                   1         1         1         0         0         0         1
#> result.8          1         1         1         1         1         1         1
#>                   1         1         1         1         1         1         1
#> result.14         1         1         1         0         1         1         0
#> result.12         1         1         1         1         1         1         1
#>           202453_s_at 213453_x_at 212975_at 202519_at 215409_at 220352_x_at
#> result.7            0           0         0         0         1           0
#>                     1           1         1         0         1           1
#> result.10           0           0         0         0         1           0
#> result.10           1           1         1         1         1           1
#>                     0           0         1         0         0           1
#> result.9            0           0         0         0         0           1
#>                     0           0         0         0         1           0
#>                     0           0         0         0         1           0
#> result.7            1           0         0         0         1           0
#>                     1           1         1         0         1           1
#>                     0           1         0         0         1           0
#> result.8            0           1         1         0         1           1
#>                     1           1         1         1         1           1
#> result.14           1           1         0         1         1           1
#> result.12           0           1         1         0         1           1
#>           209451_at 214150_x_at 212434_at 214840_at 212423_at 209798_at
#> result.7          0           0         0         0         1         0
#>                   0           1         1         1         1         0
#> result.10         0           0         0         1         1         0
#> result.10         0           1         1         1         1         1
#>                   1           1         0         1         1         0
#> result.9          1           1         0         1         1         0
#>                   0           0         0         0         1         0
#>                   0           0         0         1         1         0
#> result.7          0           1         0         0         1         0
#>                   0           0         0         1         1         1
#>                   0           1         1         1         1         1
#> result.8          0           1         1         0         1         0
#>                   0           1         1         1         1         1
#> result.14         1           0         1         1         1         0
#> result.12         0           1         1         0         1         0
#>           211061_s_at 206768_at 218042_at 218309_at 221297_at 212062_at
#> result.7            0         0         0         0         0         1
#>                     1         1         1         1         1         1
#> result.10           0         0         0         1         0         1
#> result.10           1         1         1         1         1         1
#>                     0         0         1         0         1         1
#> result.9            0         0         0         0         1         1
#>                     0         0         0         0         0         1
#>                     0         0         0         1         0         1
#> result.7            0         0         0         0         0         1
#>                     1         1         1         1         0         1
#>                     1         0         1         0         0         1
#> result.8            1         1         1         1         1         1
#>                     1         1         1         1         1         1
#> result.14           1         1         1         1         1         0
#> result.12           0         1         1         1         1         1
#>           202542_s_at 218941_at 214003_x_at 214608_s_at 204748_at 221953_s_at
#> result.7            1         0           0           0         0           0
#>                     1         0           1           1         1           0
#> result.10           1         0           0           0         1           0
#> result.10           1         1           1           1         1           0
#>                     0         0           0           1         1           0
#> result.9            0         0           0           1         1           0
#>                     1         0           0           0         1           0
#>                     1         0           0           0         1           0
#> result.7            1         0           0           0         0           1
#>                     1         1           0           0         0           0
#>                     1         0           0           1         1           1
#> result.8            1         0           1           1         1           0
#>                     1         1           1           1         1           0
#> result.14           0         1           0           1         1           1
#> result.12           1         0           1           1         1           0
#>           218580_x_at 205064_at 44563_at 206783_at 213170_at 203257_s_at
#> result.7            0         0        0         0         0           0
#>                     1         0        1         1         1           1
#> result.10           0         0        0         0         0           0
#> result.10           1         1        1         1         1           1
#>                     0         0        1         1         1           0
#> result.9            0         0        1         1         1           0
#>                     0         0        0         0         0           0
#>                     0         0        0         0         0           0
#> result.7            0         1        0         0         0           0
#>                     1         1        1         0         1           1
#>                     0         0        0         0         0           1
#> result.8            1         0        1         1         1           1
#>                     1         1        1         1         1           1
#> result.14           0         1        1         0         1           1
#> result.12           1         0        1         1         1           1
#>           203397_s_at 218895_at 202312_s_at 201787_at 204400_at 202302_s_at
#> result.7            0         1           1         0         0           0
#>                     1         1           1         1         1           1
#> result.10           0         1           1         0         0           0
#> result.10           1         1           1         1         1           1
#>                     1         0           0         1         1           1
#> result.9            1         0           0         0         1           0
#>                     0         1           1         0         0           0
#>                     0         1           1         0         0           0
#> result.7            0         0           0         0         1           0
#>                     1         1           1         1         1           0
#>                     0         1           1         0         1           1
#> result.8            1         1           1         0         1           1
#>                     1         1           1         1         1           1
#> result.14           1         0           1         1         1           1
#> result.12           1         1           1         1         1           1
#>           207070_at 217765_at 219645_at 214593_at 211868_x_at 204315_s_at
#> result.7          0         0         0         0           1           0
#>                   1         0         1         0           0           1
#> result.10         0         0         0         0           0           0
#> result.10         1         1         1         1           1           1
#>                   0         1         1         0           1           0
#> result.9          0         1         0         0           1           0
#>                   0         0         0         0           1           0
#>                   0         0         0         0           0           0
#> result.7          0         1         0         0           1           0
#>                   1         1         0         0           1           1
#>                   1         1         1         0           1           1
#> result.8          1         0         1         0           0           1
#>                   1         1         1         1           1           1
#> result.14         0         1         0         0           0           1
#> result.12         1         0         1         0           0           1
#>           210288_at 201689_s_at 214930_at 216444_at 203557_s_at 39835_at
#> result.7          0           1         0         1           0        0
#>                   1           1         1         1           1        1
#> result.10         0           1         0         1           0        0
#> result.10         1           1         1         1           1        1
#>                   0           0         1         0           1        1
#> result.9          0           0         0         0           1        0
#>                   0           1         0         1           0        0
#>                   0           1         0         1           0        0
#> result.7          0           1         0         0           1        0
#>                   1           1         1         1           0        1
#>                   0           1         0         1           1        1
#> result.8          1           1         1         1           1        1
#>                   1           1         1         1           1        1
#> result.14         1           1         1         1           1        1
#> result.12         1           1         1         1           1        1
#>           202305_s_at 220670_at 221219_s_at 214941_s_at 211665_s_at 200974_at
#> result.7            1         0           0           0           0         0
#>                     1         1           0           1           1         1
#> result.10           1         0           0           0           0         0
#> result.10           1         1           1           1           1         1
#>                     1         1           1           1           0         1
#> result.9            1         1           1           1           0         1
#>                     0         0           0           0           0         0
#>                     1         0           0           0           0         0
#> result.7            1         1           0           0           0         0
#>                     1         1           0           0           1         0
#>                     1         0           1           1           0         1
#> result.8            1         1           0           1           1         1
#>                     0         1           1           1           1         1
#> result.14           0         1           1           0           1         1
#> result.12           1         1           0           1           1         1
#>           211066_x_at 201245_s_at 42361_g_at 214304_x_at 201225_s_at 1494_f_at
#> result.7            1           1          0           1           0         0
#>                     0           1          1           1           1         1
#> result.10           1           1          0           1           0         0
#> result.10           1           1          1           1           1         1
#>                     0           0          1           0           1         1
#> result.9            0           0          1           0           0         0
#>                     1           1          0           1           0         0
#>                     1           1          0           1           0         0
#> result.7            1           1          0           0           1         0
#>                     1           1          1           1           1         1
#>                     1           1          0           1           1         0
#> result.8            1           1          1           1           1         1
#>                     1           1          1           1           0         1
#> result.14           1           1          0           1           0         1
#> result.12           0           1          1           1           1         1
#>           218653_at 206010_at 209287_s_at 204857_at 217214_s_at 216919_at
#> result.7          0         0           1         1           0         0
#>                   1         1           1         1           1         1
#> result.10         0         0           1         1           0         1
#> result.10         1         1           1         1           1         1
#>                   1         0           1         1           1         0
#> result.9          1         0           1         0           1         0
#>                   0         0           1         1           0         1
#>                   0         0           1         1           0         1
#> result.7          0         0           1         0           0         0
#>                   1         0           1         1           1         1
#>                   0         0           1         1           0         1
#> result.8          1         1           1         1           1         0
#>                   1         1           1         1           1         1
#> result.14         1         1           1         1           1         1
#> result.12         1         1           0         1           0         0
#>           218306_s_at 208328_s_at 213379_at 211370_s_at 211263_s_at 203906_at
#> result.7            0           0         0           0           1         0
#>                     1           1         1           1           1         1
#> result.10           0           0         0           0           1         0
#> result.10           1           1         1           1           1         1
#>                     0           0         0           1           0         1
#> result.9            0           0         0           1           0         1
#>                     0           0         0           1           1         0
#>                     0           0         0           0           1         0
#> result.7            0           0         0           1           1         0
#>                     1           0         1           1           1         1
#>                     0           0         1           1           1         1
#> result.8            1           1         1           1           1         0
#>                     1           1         1           1           1         1
#> result.14           1           1         1           1           1         1
#> result.12           1           1         1           1           1         0
#>           205524_s_at 212900_at 216668_at 203398_s_at 205106_at 207812_s_at
#> result.7            0         0         0           0         0           0
#>                     1         1         1           1         1           1
#> result.10           0         0         0           0         0           0
#> result.10           1         1         1           1         1           1
#>                     0         1         1           1         1           1
#> result.9            0         1         1           1         1           0
#>                     0         0         0           0         0           0
#>                     0         0         0           0         0           0
#> result.7            0         0         0           0         0           0
#>                     1         1         1           1         1           1
#>                     0         0         1           0         1           1
#> result.8            1         1         1           1         1           1
#>                     1         1         1           1         1           1
#> result.14           0         1         1           1         1           1
#> result.12           1         1         1           1         1           1
#>           211921_x_at 202926_at 206606_at  k                        rnkIndex
#> result.7            0         1         0  2  0.108880468  14.49866        1
#>                     1         1         1 10  0.016376325 254.46413        1
#> result.10           0         1         0  2  0.091279244  97.48479        1
#> result.10           1         1         1  2  0.064984472 160.66696        1
#>                     0         0         1  7  0.040942128 227.00361        1
#> result.9            0         0         1  5  0.050569465 185.49095        1
#>                     0         1         0  2  0.089521062  29.52722        2
#>                     0         1         0 10 -0.007287669 241.64255        2
#> result.7            0         1         0  4  0.046619935  75.54845        2
#>                     1         1         1  2  0.086513902  62.69761        2
#>                     1         1         1  7  0.029269549 171.27810        2
#> result.8            1         1         1  6  0.045637835 122.55207        2
#>                     1         1         1  3  0.039683388 145.01987        2
#> result.14           1         1         1  7  0.021283369 224.89984        2
#> result.12           1         1         1 10  0.022574382 222.37007        2
#>               CrowD
#> result.7        Inf
#>                 Inf
#> result.10 0.9381609
#> result.10 0.6797534
#>           0.5511496
#> result.9  0.4589268
#>                 Inf
#>                 Inf
#> result.7  0.5684247
#>           0.5230166
#>           0.4472203
#> result.8  0.3338430
#>           0.3267032
#> result.14 0.3173944
#> result.12 0.2788557
#> 
#> Slot "ParetoFront":
#> [[1]]
#>                  [,1]      [,2]
#> result.1  0.029072714  59.57509
#> result.2  0.026890329 100.89280
#> result.3  0.003859345  76.48396
#> result.4  0.027466504  20.62048
#> result.5  0.019864753  47.13439
#> result.6  0.025348479  27.85536
#> result.7  0.108880468  14.49866
#> result.8  0.006243922 195.69724
#> result.9  0.050569465 185.49095
#> result.10 0.064984472 160.66696
#> result.11 0.050805401  35.34028
#> result.12 0.004287988 119.05332
#> result.13 0.051706671  48.16061
#> result.14 0.021283369 224.89984
#> 
#> [[2]]
#>                                 
#> result.7  0.108880468  14.498664
#> result.14 0.021283369 224.899836
#> result.10 0.091279244  97.484791
#> result.10 0.064984472 160.666960
#> result.9  0.050569465 185.490954
#> result.12 0.022574382 222.370070
#> result.8  0.006243922 195.697243
#> result.1  0.066353108   1.783561
#> result.15 0.009783651 142.755291
#> result.8  0.045637835 122.552069
#> result.7  0.046619935  75.548449
#> result.3  0.058048682  36.272757
#> result.13 0.051706671  48.160610
#> result.6  0.052897230  18.820122
#> result.14 0.030435647 121.988320
#> 
#> [[3]]
#>                                 
#> result.7  0.108880468  14.498664
#>           0.016376325 254.464130
#> result.10 0.091279244  97.484791
#> result.10 0.064984472 160.666960
#> result.9  0.050569465 185.490954
#> result.12 0.022574382 222.370070
#> result.14 0.021283369 224.899836
#> result.8  0.006243922 195.697243
#>           0.086300777  46.314727
#>           0.029269549 171.278104
#> result.13 0.051706671  48.160610
#> result.7  0.046619935  75.548449
#>           0.039683388 145.019866
#> result.8  0.045637835 122.552069
#> result.1  0.066353108   1.783561
#> 
#> [[4]]
#>                                 
#> result.7   0.108880468  14.49866
#>            0.016376325 254.46413
#> result.10  0.091279244  97.48479
#> result.10  0.064984472 160.66696
#> result.9   0.050569465 185.49095
#> result.12  0.022574382 222.37007
#> result.14  0.021283369 224.89984
#>            0.089521062  29.52722
#>           -0.007287669 241.64255
#> result.7   0.046619935  75.54845
#>            0.086513902  62.69761
#>            0.014510663 217.47954
#> result.8   0.045637835 122.55207
#>            0.039683388 145.01987
#>            0.029269549 171.27810
#> 
#> [[5]]
#>                                 
#> result.7   0.108880468  14.49866
#>            0.016376325 254.46413
#> result.10  0.091279244  97.48479
#> result.10  0.064984472 160.66696
#>            0.040942128 227.00361
#> result.9   0.050569465 185.49095
#>            0.089521062  29.52722
#>           -0.007287669 241.64255
#> result.7   0.046619935  75.54845
#>            0.086513902  62.69761
#>            0.029269549 171.27810
#> result.8   0.045637835 122.55207
#>            0.039683388 145.01987
#> result.14  0.021283369 224.89984
#> result.12  0.022574382 222.37007
#> 
#>