A default call_back function that does nothing.

A default call_back function that does nothing.

callback_default (userdir = "",generation, pop_pool,
pareto, prob_matrix, current_time)

Arguments

userdir
the default directory used by galgo() to store files

generation
a number indicating the number of iterations of the galgo algorithm

pop_pool
a data.frame with the solution vectors, number of clusters and their ranking.

pareto
the solutions found by Galgo across all generations in the solution space

prob_matrix
a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns

current_time
an POSIXct object

Value

Nothing

Examples

# load example dataset

library(breastCancerTRANSBIG)
data(transbig)
Train <- transbig
rm(transbig)

#> Warning: object 'transbig' not found

expression <- Biobase::exprs(Train)
clinical <- Biobase::pData(Train)
OS <- survival::Surv(time = clinical$t.rfs, event = clinical$e.rfs)
# We will use a reduced dataset for the example
expression <- expression[sample(1:nrow(expression), 100), ]

# Now we scale the expression matrix
expression <- t(scale(t(expression)))

# Run galgo with default_callback assigned to all the hook-points

GSgalgoR::galgo(generations = 5,
population = 15,
prob_matrix = expression,
OS = OS,
start_galgo_callback = callback_default,# When Galgo is about to start.
end_galgo_callback = callback_default,  # When Galgo is about to finish.
start_gen_callback = callback_default,  # At the beginning of each iteration.
end_gen_callback = callback_default,    # At the end of each iteration.
)

#> Using CPU for computing pearson distance

#> Generation 1 Non-dominated solutions:

#>           k                      rnkIndex     CrowD
#> result.4  2 0.06285244  14.69993        1       Inf
#> result.7  8 0.02611886 242.55674        1 0.5709858
#> result.8  2 0.06223437  30.02692        1 1.2441679
#> result.9  8 0.01400487 259.18992        1       Inf
#> result.10 6 0.03488777 204.79137        1 1.4719630

#> Generation 2 Non-dominated solutions:

#>           k                      rnkIndex     CrowD
#> result.9  8 0.01400487 259.18992        1       Inf
#> result.4  2 0.09351651  95.59225        1       Inf
#> result.11 9 0.04772567 178.76986        1 1.0701305
#> result.10 6 0.03488777 204.79137        1 0.4871484
#> result.7  8 0.02611886 242.55674        1 0.4422552

#> Generation 3 Non-dominated solutions:

#>           k                      rnkIndex     CrowD
#> result.9  8 0.01400487 259.18992        1       Inf
#>           2 0.09476821  43.19902        1       Inf
#> result.11 9 0.04772567 178.76986        1 1.0341086
#> result.4  2 0.09351651  95.59225        1 1.0154582
#> result.10 6 0.03488777 204.79137        1 0.4722853
#> result.7  8 0.02611886 242.55674        1 0.4283791

#> Generation 4 Non-dominated solutions:

#>           k                      rnkIndex     CrowD
#> result.9  8 0.01400487 259.18992        1       Inf
#>           2 0.09476821  43.19902        1       Inf
#> result.11 9 0.04772567 178.76986        1 0.9599861
#>           3 0.08423763  99.69524        1 0.8287517
#> result.10 6 0.03488777 204.79137        1 0.4971586
#> result.7  8 0.02611886 242.55674        1 0.4491480
#> result.4  2 0.09351651  95.59225        1 0.3521661

#> Generation 5 Non-dominated solutions:

#>           k                     rnkIndex     CrowD
#>           2 0.10086594 107.9363        1       Inf
#>           4 0.02213126 275.5217        1       Inf
#> result.11 9 0.04772567 178.7699        1 1.0404956
#> result.10 6 0.03488777 204.7914        1 0.4648585
#> result.7  8 0.02611886 242.5567        1 0.4024423

#> NULL